FE-39: Добавил цвета
