+++ app/src/main/java/com/example/gorodbezproblem/ui/theme/Colors.kt
@@ -1,0 +1,26 @@
+package com.example.gorodbezproblem.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+class Colors {
+    companion object {
+        val YellowGreen = Color(0xFFC4DD34)
+        val Black = Color(0xFF150B0B)
+        val ThreeQuarterBlack = Color(0xBF150B0B)
+        val HalfBlack = Color(0x80150B0B)
+        val QuarterBlack = Color(0x40150B0B)
+        val Background = Color(0xFFF9F9F9)
+        val White = Color(0xFFFFFFFF)
+        val Yellow = Color(0xFFE5DE4B)
+        val Red = Color(0xFFE5624B)
+    }
+}
+
+
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
+
+val Purple40 = Color(0xFF6650a4)
+val PurpleGrey40 = Color(0xFF625b71)
+val Pink40 = Color(0xFF7D5260)
